requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.23.0
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
# torch==1.12.0+cpu
# torchaudio==0.12.0+cpu
# torchvision==0.13.0+cpu
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.9
